requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
python-multipart==0.0.9
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
rank-bm25==0.2.2
sentence-transformers==3.0.1
langchain-chroma==0.1.1
faiss-cpu==1.8.0.post1
qdrant-client==1.9.1
pypdf==4.3.1
docx2txt==0.8
markdown==3.6
pdfminer.six==20221105
python-dotenv==1.0.1
streamlit==1.37.1
ragas==0.1.19
pandas==2.2.2
matplotlib==3.9.2
pyyaml==6.0.2
tqdm==4.66.4
requests==2.32.3
httpx==0.27.0
pytest==8.3.2
cohere==5.5.8
FlagEmbedding==1.2.10
